producer/requirements.txt
confluent-kafka==2.3.0
pandas
PyYAML>=5.1.2
fastavro
requests